+++ src/JsonApi/Response/FetchResponse.php
@@ -58,7 +58,7 @@ class FetchResponse extends AbstractResponse
      * data (as described above)."
      *
      * @param \WoohooLabs\Yin\JsonApi\Transformer\AbstractErrorDocument $document
-     * @param array $errors
+     * @param \WoohooLabs\Yin\JsonApi\Schema\Error[] $errors
      * @return \Psr\Http\Message\ResponseInterface $response
      */
     public function notFound(AbstractErrorDocument $document, array $errors = [])