+++ tests/JsonApi/Exception/InclusionNotSupportedTest.php
@@ -1,0 +1,19 @@
+<?php
+namespace WoohooLabsTest\Yin\JsonApi\Schema;
+
+use PHPUnit_Framework_TestCase;
+use WoohooLabs\Yin\JsonApi\Exception\InclusionNotSupported;
+
+class InclusionNotSupportedTest extends PHPUnit_Framework_TestCase
+{
+    public function testGetMessage()
+    {
+        $exception = $this->createException();
+        $this->assertEquals("Inclusion is not supported!", $exception->getMessage());
+    }
+
+    private function createException()
+    {
+        return new InclusionNotSupported();
+    }
+}