+++ tests/JsonApi/Exception/InclusionUnrecognizedTest.php
@@ -1,0 +1,21 @@
+<?php
+namespace WoohooLabsTest\Yin\JsonApi\Schema;
+
+use PHPUnit_Framework_TestCase;
+use WoohooLabs\Yin\JsonApi\Exception\InclusionUnrecognized;
+
+class InclusionUnrecognizedTest extends PHPUnit_Framework_TestCase
+{
+    public function testGetIncludes()
+    {
+        $includes = ["a", "b", "c"];
+
+        $exception = $this->createException($includes);
+        $this->assertEquals($includes, $exception->getIncludes());
+    }
+
+    private function createException(array $includes)
+    {
+        return new InclusionUnrecognized($includes);
+    }
+}